requirement.txt
pandas==1.5.2
webdriver-manager==3.8.5
python-dotenv==0.21.0
requests==2.28.2
beautifulsoup4==4.11.1
selenium==4.7.2
openpyxl==3.0.10
python-dotenv==0.21.0
pytest==7.2.1